Added a universal GUI handler
