+++ pycos/src/system/sysdata.py
@@ -1,0 +1,2 @@
+pycos_version = "0.0.2"
+reset_causes = ["PWRON_RESET", "HARD_RESET", "WDT_RESET", "DEEPSLEEP_RESET", "SOFT_RESET"]
